Add support for operating on calculation result
Currently, if the user types 2 + 3 = and gets an answer of 5, they will have to manually enter 5 to use it in a calculation.

Update the calculator so the user can press an operator right after they have pressed the equal sign.

This will begin a new calculation where the last result is the first operand.

For example, pressing 2 + 3 =, followed by * 6, will display 5 * 6.